requirements.txt
Django>=2.2.1
django-apptemplates>=1.4
django-filter>=2.1.0
djangorestframework>=3.9.4
Pillow>=6.0.0
pkg-resources>=0.0.0
pytz>=2019.1
sqlparse>=0.3.0
